Update Ember-CLI and add tests/dummy/public
I updated to Ember-CLI 2.17.1 and ran the related blueprints.

I also, I think have a partial solution for https://github.com/ivanvotti/ember-svg-jar/issues/29

I added a check in src/index.js like:

```js
if(isAddon) {
  defaultOpts.sourceDirs.push('tests/dummy/public');
  }
```

So we should be automatically adding `tests/dummy/public` to the directories we look through, if you are inside an addon.

I tried linking locally to an addon to test, but when I ran `ember s` the addon could not access http://localhost:4200/ember-svg-jar/index.html so I am not 100% sure if this works or not, but I think it should.