Better addon support
This change allows ember-svg-jar to be used by addons. [Recent work](https://github.com/ivanvotti/ember-svg-jar/issues/29) also mentioned making it work in addons, but that only makes it work within an addon's own dummy app, when consumed as a `devDependency`.

This change makes ember-svg-jar work as a `dependency`. Multiple addons and/or the top-level app can all use it simultaneously without stomping each other (with the one caveat that the namespace of SVG asset ids is flat, so `{{svg-jar "something"}}` has a consistency meaning no matter where it appears.

I only implemented this for the default, inline strategy. It would be possible in the future to do for the symbol strategy.