Use `uniqueId` for recurring `title` and `desc`
The current implementation uses `guidFor` this will cache ids and reuse
them if the same value is passed to `guidFor`, see
https://github.com/emberjs/ember.js/blob/4e3300bdfe75da14f9714b6b1539dbd1612c5af2/packages/%40ember/-internals/utils/lib/guid.ts#L97

The ids in a HTML document should be unique. However if a `title` or
`desc` is reused, for example you have a recurring element, then
`svg-jar` generates ids which match for the recurring element.

This commit uses the `uniqueId` function from the `unique-id` helper
to generate a unique id regardless of the `svg-jar` `title` or `desc`
matching previous `svg-jar`.

Unfortunately `uniqueId` is not exported so it is copy and pasted.
In the future it will be possible to import it, see
https://github.com/emberjs/ember.js/issues/20165

Fixes https://github.com/voltidev/ember-svg-jar/issues/243